Use sigaction for interrupt handler
Prior to this commit, we were using signal to register our interrupt handler. This resulted in potential "I/O possible" termination errors, as sigaction is the preferred way of registering signal handlers currently. This commit changes joystick-commander to use sigaction instead of signal, minimizing the risk of the previous errors.