Clear out exponential filter averages on disable
Prior to this commit, the exponential filter of the inputs
(triggers and joystick) were filtered so the modules received a smooth
series of commands, but the filters were not cleared after a disable,
causing the modules to actuate on old command values when re-enabling.
This commit clears the filters when control is not enabled.