Make initialization of `joystick_data` explicit
Prior to this commit the initialization of the static global variable `joystick_data` was initialized ambiguously and what looked like erroneously.

This commit represents a revision explicit initialization instead.

Motivation for this change was a support ticket raised with an email to `help@polysync.io`

Note to reviewers: I validated that after this change the joystick commander still behaved as _I_ expected on the car but would HIGHLY recommened that reviewers do the same.

Thanks!